requirements.txt
BeautifulSoup==3.2.1
argparse==1.2.1
distribute==0.6.24
fusepy==2.0.2
soupselect==0.0.0